refactor: move WSEditorCommandGroupContenet and sub-components to own dir
