+++ src/web/src/views/workspace/commandGroupContent/index.tsx
@@ -1,0 +1,1 @@
+export { default, DecodeResponseCommandGroup } from "./WSEditorCommandGroupContent";
